requirements.txt
Flask
gunicorn
googletrans==4.0.0-rc1
psycopg2-binary
Werkzeug